make handling of unit deletions more precise
Whenever we propagate unit clauses (that are part of the formula)
we map the clause to the unit that was propagated.

Now if a clause is deleted, we check if there was a propagation where this
clause was the reason. If so, we reset the assignment to the level immediately
prior to propagating that unit.