skip definitely satisfied premises
This adds a sound approximation to skip adding satisfied premises / (pseudo)
tautologies.

In other checkers and with `rate --skip-deletions`, premises are skipped if they
are satisfied by the current assignment.

If we always perform deletions the problem arises that satisfied clauses can be
falsfied later in the proof, if the reason for a satisfied literal is deleted.
Here, whenever we assign a literal in forward mode, we compute its minimal
lifetime, that is the proof step where it first may by falsified. The minimal
lifetime of a clause is the minimum of its deletion time and all minimal
lifetimes of its falsified literals (except for the satisfied one).